hawkbit: raise a WARN if a action id is not in env
For long time, thjere was a race condition with Hawkbit: if after a
restart, the operator changed on the server the deploy, SWUpdate
committed an update that wasn't executed. To avoid this, the action id
of the update that was executed is now stored in env, but to be
compatible with the past, the old behavior is accepted as well.

This raises a warning in case the old behavior is chosen to advise that
the answer to Hawkbit could be wrong.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Reviewed-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
(cherry picked from commit d0cfc4ec27eb6ccb25612a6f325982db7f6fdd3d)
Upstream-Status: Backport